move App.js to src directory and update imports
