+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import '../App.css';
-import MainWindow from './MainWindow';
+import './App.css';
+import MainWindow from './components/MainWindow';
