+++ client/src/pages/Start.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Start = () => {
+  return (
+    <h1>ASTEROIDS</h1>
+  )
+};
+
+export default Start;
